Move indentex types to a separate source file
